requirements.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.3
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
factory-boy==2.11.1
Faker==1.0.5
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.7.7
pycodestyle==2.5.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.2
whitenoise==4.1.2
